Implement assignment and composite assignment (#95)
Implement assignment (i.e. `a = 2`) and composite assignment (i.e. `a += 1`, `a &= 3` , etc) operators.

Refactor the parser a little bit to make it possible to share the checking code between regular binary operations and their composite assignment variants.

Before the parser had `Expr BuildBinaryAddition`, that performed the required checks, promoted operands and created the result expression. For composite assignment, we  need additional checks and don't want to modify the `lhs` expression -- it has to stay "as is", because we're going to modify it during the evaluation. So now we have `Type PrepareBinaryAddition`, which performs the checks, promotes the RHS operand (if needed) and return the promoted type of the binary operation. Then for composite assignments we do additional checks via `PrepareCompositeAssignment` and create the final expression.